More work on organizers datalist extraction
